+++ CHANGELOG.md
@@ -12,6 +12,7 @@ used to track changes in this project.
 ### Changed
 - [#42](../../issues/42): Change all use case Actions to require a string ID instead of UuidInterface
 - [#21](../../issues/21): Cart price rules carried over and linked to an Order (DB migration)
+- [#40](../../issues/40): Change ProductRepository::getRelatedProductsByIds to not require $tagIds
 
 ### DB Migration
 - TBD